requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.41.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=0.21.0
websockets==12.0
pydantic==2.5.3
aiohttp>=3.8.0
lxml==4.9.3
httpx>=0.24.0
